loading of spinners in NhisClaimFormActivity
Created classes for 'type of service', 'outcome', 'attendance' ,
'procedure', investigation'  and 'medicines' which are all properties of
the Claim. Created dummy data and entered it into the database for this
properties. In view of this I modified the DataClass method updating the
dataversion to 15 and creating an upgradeToVersion15 method. I modified
the NhisClaimFormActivity to fill spinners.